+++ src/app/cv/builder/components/preview/index.tsx
@@ -7,22 +7,34 @@ import ResumeProjects from "../resume-section-displays/resume-projects-display";
 import ResumeSkills from "../resume-section-displays/resume-skills-display";
 import ResumeSummary from "../resume-section-displays/resume-summary-display";
 import styles from "./index.module.scss";
+import { useActiveSections } from "../../hooks";
 
 export default function Preview() {
+  const { storeGetActiveSections } = useActiveSections();
+  const activeSectionNames = storeGetActiveSections();
+
   return (
     <div className={styles["preview"]}>
-      <ResumeBasicInfo />
+      {activeSectionNames.includes("basicInfo") ? <ResumeBasicInfo /> : null}
       <div className={styles.sections}>
         <Row>
           <Col span={14}>
-            <ResumeSummary />
-            <ResumeExperience />
-            <ResumeProjects />
+            {activeSectionNames.includes("summary") ? <ResumeSummary /> : null}
+            {activeSectionNames.includes("experience") ? (
+              <ResumeExperience />
+            ) : null}
+            {activeSectionNames.includes("projects") ? (
+              <ResumeProjects />
+            ) : null}
           </Col>
           <Col span={10}>
-            <ResumeSkills />
-            <ResumeEducation />
-            <ResumeLanguages />
+            {activeSectionNames.includes("skills") ? <ResumeSkills /> : null}
+            {activeSectionNames.includes("education") ? (
+              <ResumeEducation />
+            ) : null}
+            {activeSectionNames.includes("languages") ? (
+              <ResumeLanguages />
+            ) : null}
           </Col>
         </Row>
       </div>